added color and moving train at end

+++ db/migrate/20170912003914_create_ratings.rb
@@ -1,7 +1,7 @@
 class CreateRatings < ActiveRecord::Migration[5.0]
   def change
     create_table :ratings do |t|
-      t.integer :star_rating, :default => nil
+      t.float :star_rating, :default => nil
       t.boolean :fav, :default => nil
       t.belongs_to :user, index: true
       t.belongs_to :recipe, index: true
